Turn StaticArrays into an Extension Pacakge
omg so fast!
no requires on v1.8 extension


typo


fix imports


fix all tests


fix the backwards compat with no requires


add missing ones


tag


overload valtype


add partials


Update ext/ForwardDiffStaticArraysExt.jl

a few more imports


finish the migration
