+++ src/ForwardDiff.jl
@@ -2,7 +2,6 @@ module ForwardDiff
 
 using DiffRules, DiffResults
 using DiffResults: DiffResult, MutableDiffResult, ImmutableDiffResult
-using StaticArrays
 if VERSION >= v"1.6"
     using Preferences
 end
@@ -24,6 +23,10 @@ include("derivative.jl")
 include("gradient.jl")
 include("jacobian.jl")
 include("hessian.jl")
+
+if !isdefined(Base, :get_extension)
+    include("../ext/ForwardDiffStaticArraysExt.jl")
+end
 
 export DiffResults
